requirements.txt
langchain_core>=0.1.14
langgraph>=0.0.19
langchain_openai>=0.0.6
streamlit>=1.31.0
python-dotenv>=1.0.0

